FIX: compiler warning double to int conversion
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@1391 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Base/cmtkUniformVolumePainter.cxx
@@ -113,8 +113,8 @@ cmtk::UniformVolumePainter::DrawBox
       // nothing to do - already indexed
       for ( int dim = 0; dim < 3; ++dim )
 	{
-	indexFrom[dim] = boxFrom[dim];
-	indexTo[dim] = boxTo[dim];
+	indexFrom[dim] = static_cast<int>( boxFrom[dim] );
+	indexTo[dim] = static_cast<int>( boxTo[dim] );
 	}
       break;
     }
